Create new ReviewHomeWrapper to keep most of the logic for filtering and show the ReviewHeader (linking to application) and keep ReviewHome simpler
