+++ src/components/Review/ReviewContainer.tsx
@@ -1,19 +1,15 @@
 import React from 'react'
-import { Container, Header, Icon, Label, Segment } from 'semantic-ui-react'
+import { Container, Header, Icon, Label } from 'semantic-ui-react'
 import { useRouter } from '../../utils/hooks/useRouter'
 import { ApplicationDetails } from '../../utils/types'
 import strings from '../../utils/constants'
-import { Link } from 'react-router-dom'
 
 export interface ReviewContainerProps {
   application: ApplicationDetails
 }
 
 const ReviewContainer: React.FC<ReviewContainerProps> = ({ application, children }) => {
-  const {
-    query: { serialNumber },
-    push,
-  } = useRouter()
+  const { push } = useRouter()
   const { template, name, org } = application
   return (
     <Container id="review-area">
@@ -32,14 +28,6 @@ const ReviewContainer: React.FC<ReviewContainerProps> = ({ application, children
           {org?.name || strings.TITLE_NO_ORGANISATION}
         </Header>
       </div>
-      <Segment className="review-application-segment" basic textAlign="center">
-        <Header
-          as="h2"
-          className="clickable"
-          onClick={() => push(`/application/${serialNumber}`)}
-          content={name}
-        />
-      </Segment>
       {children}
     </Container>
   )
